Fix UB caused by using withUnsafeMutablePointer
Before this commit there was UB in findObject function.
When certificates were enumerated, we took CKA_ID of each cert
and used it to find corresponding private key.
Misuse of "withUnsafeMutablePointer" cause to inner pointer to be used
outside of closure and leads to UB - sometimes,
when CKA_ID length was small, memory was reused by something else.

Resolve: SHIFTIOS-54