Fix keyboard avoidance issues on PinInputView
The concept in SwiftUI layout system is that views can have a fixed size
or a flexible size. Spacers, Colors, ScrollViews, etc
have flexible sizes, so they tend to expand to occupy all the space
available. When the keyboard is showing, there is a safe area applied on
the container. The height of the container will decrease.
The ignoresSafeArea modifier is typically applied on views that have a
flexible height. If ignoresSafeArea is applied, it will makes the
ScrollView extend its height to ignore the keyboard bottom safe area.

Refer: #41