Add keychain group
Launch app on macOS may result in Error -34018 errSecMissingEntitlement.
The immediate cause of this error is that you’re trying to do something
with the Security framework that requires an entitlement and your app
was not signed with that entitlement. The entitlement in question is
usually the application-identifier entitlement, which is the primary
mechanism by which the Security framework identifies your app.

Resolve: SHIFTIOS-104